+++ src/main/kotlin/com/github/aarcangeli/ideaclangformat/ClangFormatConfig.kt
@@ -14,6 +14,8 @@ class ClangFormatConfig : SimplePersistentStateComponent<ClangFormatConfig.State
   class State : BaseState() {
     var enabled by property(true)
 
+    var formatOnSave by property(false)
+
     /// The path to the clang-format executable.
     /// If null, the plugin will try to find it in the PATH.
     var customPath by string()