+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Detect indentation style and column limit from `.clang-format` file
 - Format code using `clang-format` binary
 - Settings page to configure `clang-format` binary path
-- Format on save option
+- On Rider, format code on save (other IDEs already have this feature)
 
 [Unreleased]: https://github.com/aarcangeli/idea-clang-format/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/aarcangeli/idea-clang-format/commits/v1.0.0